+++ app/src/main/java/com/wallapp/activities/IntroActivity.java
@@ -1,0 +1,41 @@
+package com.wallapp.activities;
+
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.os.Bundle;
+import android.preference.PreferenceManager;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AppCompatActivity;
+
+import com.wallapp.MainActivity;
+
+public class IntroActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        if (ContextCompat.checkSelfPermission(IntroActivity.this,
+                Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(IntroActivity.this,
+                    new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 200);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        if (requestCode == 200) {
+            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                startActivity(new Intent(getBaseContext(), MainActivity.class));
+                PreferenceManager.getDefaultSharedPreferences(getBaseContext())
+                        .edit().putBoolean("firstStart", false).apply();
+                finish();
+            }
+        }
+    }
+}