+++ app/src/main/java/com/wallapp/utils/CustomConstants.java
@@ -1,8 +1,8 @@
 package com.wallapp.utils;
 
 public interface CustomConstants {
-    String UNO_BASE_URL = "https://unsplash.it/";
-    String DOS_BASE_URL = "https://source.unsplash.com/";
+    String ALPHA_BASE_URL = "https://unsplash.it/";
+    String GAMMA_BASE_URL = "https://source.unsplash.com/";
     String BING_BASE_URL = "https://bing.com/";
     String BING_DAILY_URL = BING_BASE_URL + "HPImageArchive.aspx?format=js&n=7&idx=0";
 }