+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Version 1.6.3
+
+  - Added edit feature
+  - Settings tweaks
+  - Removed intro (that's boring, anyway)
+
 ## Version 1.6.1
 
   - Updated libraries